formatting and moved replica config creation to server
