+++ mii/deployment.py
@@ -7,14 +7,11 @@ import string
 import os
 import mii
 
-from deepspeed.launcher.runner import fetch_hostfile
-
 from .constants import DeploymentType, MII_MODEL_PATH_DEFAULT, MODEL_PROVIDER_MAP
 from .logging import logger
 from .utils import get_task_name, get_provider_name
 from .models.score import create_score_file
 from .models import load_models
-from .config import ReplicaConfig, LoadBalancerConfig
 
 
 def deploy(task,
@@ -107,29 +104,6 @@ def deploy(task,
     elif model_path is None and deployment_type == DeploymentType.AML:
         model_path = "model"
 
-    # add fields for replica deployment
-    replica_pool = _allocate_processes(mii_config.hostfile,
-                                       mii_config.tensor_parallel,
-                                       mii_config.replica_num)
-    print(replica_pool)
-    replica_configs = []
-    for i, (hostname, gpu_indices) in enumerate(replica_pool):
-        # Reserver port for a LB proxy when replication is enabled
-        port_offset = 1
-        base_port = mii_config.port_number + i * mii_config.tensor_parallel + port_offset
-        tensor_parallel_ports = list(
-            range(base_port,
-                  base_port + mii_config.tensor_parallel))
-        torch_dist_port = mii_config.torch_dist_port + i
-        replica_configs.append(
-            ReplicaConfig(hostname=hostname,
-                          tensor_parallel_ports=tensor_parallel_ports,
-                          torch_dist_port=torch_dist_port,
-                          gpu_indices=gpu_indices))
-
-    lb_config = LoadBalancerConfig(port=mii_config.port_number,
-                                   replica_configs=replica_configs)
-
     if deployment_type != DeploymentType.NON_PERSISTENT:
         create_score_file(deployment_name=deployment_name,
                           deployment_type=deployment_type,
@@ -139,8 +113,7 @@ def deploy(task,
                           ds_zero=enable_zero,
                           ds_config=ds_config,
                           mii_config=mii_config,
-                          model_path=model_path,
-                          lb_config=lb_config)
+                          model_path=model_path)
 
     if deployment_type == DeploymentType.AML:
         _deploy_aml(deployment_name=deployment_name,
@@ -185,41 +158,3 @@ def _deploy_aml(deployment_name, instance_type, version):
         f"AML deployment assets at {mii.aml_related.utils.aml_output_path(deployment_name)}"
     )
     print("Please run 'deploy.sh' to bring your deployment online")
-
-
-def _allocate_processes(hostfile_path, tensor_parallel, num_replicas):
-    resource_pool = fetch_hostfile(hostfile_path)
-    assert resource_pool is not None and len(
-        resource_pool) > 0, f'No hosts found in {hostfile_path}'
-
-    replica_pool = []
-    allocated_num = 0
-    for host, slots in resource_pool.items():
-        available_on_host = slots
-        while True:  #available_on_host >= tensor_parallel:
-            if allocated_num >= num_replicas:
-                break
-            # TODO: Skip for AML / move to deployment time
-            #if slots < tensor_parallel:
-            #    raise ValueError(
-            #        f'Host {host} has {slots} slot(s), but {tensor_parallel} slot(s) are required'
-            #    )
-
-            allocated_num_on_host = slots - available_on_host
-            replica_pool.append(
-                (host,
-                 [
-                     i for i in range(allocated_num_on_host,
-                                      allocated_num_on_host + tensor_parallel)
-                 ]))
-            allocated_num += 1
-
-            available_on_host -= tensor_parallel
-
-    # TODO: this check needs to happen when we start the server for AML
-    #if allocated_num < num_replicas:
-    #    raise ValueError(
-    #        f'No sufficient GPUs for {num_replicas} replica(s), only {allocated_num} replica(s) can be deployed'
-    #    )
-
-    return replica_pool
